+++ src/components/Product/Pipelines/index.js
@@ -2,16 +2,16 @@ import CloudinaryImage from 'components/CloudinaryImage'
 import { graphql, useStaticQuery } from 'gatsby'
 import React, { useMemo } from 'react'
 import Link from 'components/Link'
-import Layout from '../Layout'
+import Layout from '../../Layout'
 import { SEO } from 'components/seo'
 import { AnimatePresence, motion } from 'framer-motion'
 import { IconSearch, IconPlug, IconArrowRightDown } from '@posthog/icons'
 import Fuse from 'fuse.js'
 import TeamMembers from '../TeamMembers'
 import SideModal from 'components/Modal/SideModal'
-import Questions from '../Product/Questions'
+import Questions from '../Questions'
 import { DocLinks } from 'components/Products/DocsLinks'
-import { docsMenu } from '../../navs'
+import { docsMenu } from '../../../navs'
 import { MDXProvider } from '@mdx-js/react'
 import { VsCompetitor } from 'components/Products/Competitor'
 import { VsPostHog } from 'components/Products/Competitor/VsPostHog'
